[Feature] myWaitingPosition 구현 (#68)
