+++ src/main/java/org/example/catch_line/history/controller/HistoryController.java
@@ -2,22 +2,22 @@ package org.example.catch_line.history.controller;
 
 import java.util.List;
 
+import org.example.catch_line.booking.reservation.service.ReservationService;
+import org.example.catch_line.booking.waiting.service.WaitingService;
 import org.example.catch_line.common.SessionUtils;
-import org.example.catch_line.common.constant.SessionConst;
 import org.example.catch_line.common.constant.Status;
+import org.example.catch_line.exception.booking.HistoryException;
 import org.example.catch_line.history.model.dto.HistoryResponse;
 import org.example.catch_line.history.service.HistoryService;
-import org.example.catch_line.reservation.service.ReservationService;
-import org.example.catch_line.waiting.service.WaitingService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
-
 
 @Controller
 @RequiredArgsConstructor
@@ -34,15 +34,66 @@ public class HistoryController {
 		@RequestParam(defaultValue = "SCHEDULED") Status status
 	) {
 		Long memberId = SessionUtils.getMemberId(session);
-		List<HistoryResponse> allHistory = historyService.getAllHistory(memberId,status);
+		List<HistoryResponse> allHistory = historyService.getAllHistory(memberId, status);
+		session.setAttribute("historyList", allHistory);
 		model.addAttribute("history", allHistory);
 
 		return "history/history";
 	}
 
-	@GetMapping("/history/waitings/{waitingId}")
-	public String getHistory() {
-		return "history/detail";
+	@GetMapping("/history/waiting/{waitingId}")
+	public String getWaitingDetail(@PathVariable Long waitingId, HttpSession session, Model model) {
+
+		List<HistoryResponse> historyList = (List<HistoryResponse>)session.getAttribute("historyList");
+
+		if (historyList != null) {
+			try {
+				HistoryResponse historyResponse = historyService.findWaitingDetailById(historyList, waitingId);
+				model.addAttribute("historyResponse", historyResponse);
+				return "history/waitingDetail";
+			} catch (IllegalArgumentException e) {
+				// 오류 처리: 메시지 표시 또는 로그 기록 등
+			}
+		}
+
+		return "redirect:/history";
+	}
+
+	@GetMapping("/history/reservation/{reservationId}")
+	public String getReservationDetail(
+		@PathVariable Long reservationId,
+		Model model,
+		HttpSession session
+	) {
+		List<HistoryResponse> historyList = (List<HistoryResponse>)session.getAttribute("historyList");
+
+		if (historyList != null) {
+			try {
+				HistoryResponse historyResponse = historyService.findReservationDetailById(historyList, reservationId);
+				model.addAttribute("historyResponse", historyResponse);
+				return "history/reservationDetail";
+			} catch (HistoryException e) {
+				model.addAttribute("errorMessage", "지금은 상세정보를 조회할 수 없습니다");
+				return "error";
+			}
+		}
+		return "redirect:/history";
+	}
+
+	@PostMapping("/history/reservation/{reservationId}")
+	public String deleteReservation(@PathVariable Long reservationId) {
+
+		reservationService.cancelReservation(reservationId);
+
+		return "redirect:/history";
+	}
+
+	@PostMapping("/history/waiting/{waitingId}")
+	public String deleteWaiting(@PathVariable Long waitingId) {
+
+		waitingService.cancelWaiting(waitingId);
+
+		return "redirect:/history";
 	}
 
 }
