--- src/main/java/org/example/catch_line/reservation/exception/IdException.java
@@ -1,7 +1,0 @@
-package org.example.catch_line.reservation.exception;
-
-public class IdException extends RuntimeException{
-	public IdException() {
-		super("올바른 경로로 접속해주세요");
-	}
-}
