+++ src/main/resources/templates/history/history.html
@@ -4,19 +4,68 @@
     <meta charset="UTF-8">
     <title>History</title>
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
+    <style>
+        /* 주황색 톤을 적용한 btn-orange 스타일 */
+        .btn-orange {
+            background-color: orange;
+            color: white;
+            border: none;
+            transition: background-color 0.3s ease;
+        }
+
+        .btn-orange:hover {
+            background-color: darkorange;
+            color: #fff;
+        }
+
+        table th, table td {
+            vertical-align: middle;
+        }
+
+        thead th {
+            background-color: #f8f9fa;
+            font-weight: bold;
+            text-align: center;
+        }
+
+        tbody tr:hover {
+            background-color: #ffebcc;
+            cursor: pointer;
+        }
+
+        /* Center align text in the table */
+        table th, table td {
+            text-align: center;
+        }
+
+        /* Add spacing between the buttons */
+        .btn-group {
+            display: flex;
+            justify-content: space-between;
+        }
+
+        .btn-group .col {
+            flex: 1;
+            display: flex;
+            justify-content: center;
+        }
+    </style>
 </head>
 <body>
 <div class="container mt-5">
     <h1 class="mb-4">History</h1>
-    <div class="row mb-4">
+    <div class="row mb-4 btn-group">
         <div class="col">
-            <button type="button" class="btn btn-primary btn-lg" th:onclick="|location.href='@{/history(status=SCHEDULED)}'|">예정</button>
+            <button type="button" class="btn btn-orange btn-lg" th:onclick="|location.href='@{/history(status=SCHEDULED)}'|">예정</button>
         </div>
         <div class="col">
-            <button type="button" class="btn btn-primary btn-lg" th:onclick="|location.href='@{/history(status=COMPLETED)}'|">완료</button>
+            <button type="button" class="btn btn-orange btn-lg" th:onclick="|location.href='@{/history(status=COMPLETED)}'|">완료</button>
         </div>
         <div class="col">
-            <button type="button" class="btn btn-primary btn-lg" th:onclick="|location.href='@{/history(status=CANCELED)}'|">취소</button>
+            <button type="button" class="btn btn-orange btn-lg" th:onclick="|location.href='@{/history(status=CANCELED)}'|">취소</button>
+        </div>
+        <div class="col">
+            <button type="button" class="btn btn-orange btn-lg" onclick="refreshPage()">새로고침</button>
         </div>
     </div>
     <table class="table table-striped">
@@ -30,10 +79,15 @@
             <th>Modified At</th>
             <th>Waiting Type / Reservation Date</th>
             <th>WaitingRegistrationId</th>
+            <th>MyWaitingPosition</th>
         </tr>
         </thead>
         <tbody>
-        <tr th:each="history : ${history}">
+        <tr th:each="history : ${history}"
+            th:data-service-type="${history.serviceType}"
+            th:data-waiting-id="${history.waitingId}"
+            th:data-reservation-id="${history.reservationId}"
+            onclick="redirectBasedOnServiceType(this)">
             <td th:text="${history.restaurantName}"></td>
             <td th:text="${history.serviceType}"></td>
             <td th:text="${history.memberCount}"></td>
@@ -42,6 +96,7 @@
             <td th:text="${history.modifiedAt != null ? #temporals.format(history.modifiedAt, 'yyyy-MM-dd HH:mm:ss') : 'N/A'}"></td>
             <td th:text="${history.serviceType == T(org.example.catch_line.restaurant.model.entity.constant.ServiceType).WAITING ? history.waitingType : (history.reservationDate != null ? #temporals.format(history.reservationDate, 'yyyy-MM-dd HH:mm:ss') : 'N/A')}"></td>
             <td th:text="${history.serviceType == T(org.example.catch_line.restaurant.model.entity.constant.ServiceType).WAITING ? history.waitingRegistrationId : 1}"></td>
+            <td th:text="${history.serviceType == T(org.example.catch_line.restaurant.model.entity.constant.ServiceType).WAITING ? history.myWaitingPosition : 1}"></td>
         </tr>
         </tbody>
     </table>
@@ -50,5 +105,22 @@
 <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+<script>
+    function refreshPage() {
+        location.reload();
+    }
+
+    function redirectBasedOnServiceType(row) {
+        const serviceType = row.getAttribute('data-service-type');
+        const waitingId = row.getAttribute('data-waiting-id');
+        const reservationId = row.getAttribute('data-reservation-id');
+
+        if (serviceType === 'WAITING') {
+            location.href = `/history/waiting/${waitingId}`;
+        } else {
+            location.href = `/history/reservation/${reservationId}`;
+        }
+    }
+</script>
 </body>
 </html>
