+++ src/main/resources/templates/history/waitingDetail.html
@@ -1,0 +1,113 @@
+<!doctype html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Waiting Details</title>
+    <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
+    <style>
+        /* 주황색 톤을 적용한 btn-danger 스타일 */
+        .btn-danger {
+            background-color: orange;
+            border-color: #ff6347;
+            color: white;
+            transition: background-color 0.3s ease;
+        }
+
+        .btn-danger:hover {
+            background-color: orange;
+            border-color: white;
+        }
+
+        .card-header {
+            background-color: orange;
+            color: #fff;
+        }
+
+        .card-body {
+            background-color: #fff8f0;
+        }
+
+        .card-body h5 {
+            color: orange;
+            font-weight: bold;
+        }
+
+        .card {
+            border-color: orange;
+        }
+
+        /* 텍스트 스타일 */
+        .card-body p {
+            font-size: 16px;
+        }
+
+        .card-body p strong {
+            color: #333;
+        }
+    </style>
+</head>
+<body>
+<div class="container mt-5">
+    <h1 class="mb-4">Waiting Details</h1>
+
+    <div class="card">
+        <div class="card-header">
+            <!-- 식당 이름을 동적으로 설정 -->
+            <h3 th:text="${historyResponse.restaurantName}">Restaurant Name</h3>
+        </div>
+        <div class="card-body">
+            <div class="row mb-3">
+                <div class="col-md-6">
+                    <h5>대기정보</h5>
+                    <!-- 접수 일시 -->
+                    <p><strong>접수 일시:</strong> <span th:text="${#temporals.format(historyResponse.createdAt, 'yyyy-MM-dd HH:mm:ss')}">2023-08-09 12:34:56</span></p>
+                    <!-- 인원수 -->
+                    <p><strong>인원수:</strong> <span th:text="${historyResponse.memberCount}">4</span></p>
+                    <!-- 이용 상태 -->
+                    <p><strong>이용상태:</strong> <span th:text="${historyResponse.status}">대기중</span></p>
+                </div>
+                <div class="col-md-6">
+                    <h5>추가 정보</h5>
+                    <!-- 대기 위치 -->
+                    <p><strong>대기 순서:</strong> <span th:text="${historyResponse.myWaitingPosition}">3</span></p>
+                    <!-- 대기 등록 ID -->
+                    <p><strong>대기 번호:</strong> <span id="waitingRegistrationId" th:text="${historyResponse.waitingRegistrationId}">123456</span></p>
+                    <p><strong>대기 ID:</strong> <span id="waitingId" th:text="${historyResponse.waitingId}">123456</span></p>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col text-right">
+                    <!-- 대기 취소 버튼 -->
+                    <button type="button" class="btn btn-danger" onclick="cancelWaiting()">취소하기</button>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+<script>
+    function cancelWaiting() {
+        const waitingId = document.getElementById('waitingId').innerText;
+        const url = `/history/waiting/${waitingId}`;
+
+        fetch(url, {
+            method: 'POST'
+        })
+            .then(response => {
+                if (response.ok) {
+                    alert('대기가 취소되었습니다.');
+                    window.location.href = '/history'; // 취소 후 history 페이지로 리디렉션
+                } else {
+                    alert('대기 취소에 실패했습니다.');
+                }
+            })
+            .catch(error => {
+                console.error('Error:', error);
+                alert('대기 취소에 실패했습니다.');
+            });
+    }
+</script>
+</body>
+</html>
